api_library/openai☑/requirements.txt
openai
tiktoken
python-dotenv
urllib3
pydantic